refactor: retarget neutral event bus owner
